Add files via upload
In this model, a linear regression model will be used to predict the value of a soccer player based on their statistics from 2018-2022. For this model, a dataset containing various features such as passes, assists, goals, the league in which they play, among other characteristics, will be used, which are very useful when creating the model.
The common phases for Machine Learning projects will be carried out, from the EDA process to model creation and evaluation.